android: Fix misc compiler errors for GCC 5.4+
I'm preparing an update of our custom toolchains (trying both
with 5.4 and 6.2 at the moment), and using them reveals a lot
of minor compilation bugs:

- Most of them come from the fact that GoogleTest's EXPECT_EQ()
  is much stricter about the types of integers being compared
  (i.e. 'signed vs unsigned' and 'int vs size_t'), so fix the
  unit tests.

  (this patch is only a start, there are a _lot_ more unit-tests
  to fix).

- android/qemu1/util/aes.c: Move static variable declaration into
  conditional code block to avoid error about an unused variable.

- android/qemu1/include/qemu/osdep.h: Avoid redefining 'inline'
  as a macro when included from a C++ source file.

More fixes will follow soon.

+++ android/android-emu/android/base/system/Win32UnicodeString_unittest.cpp
@@ -95,7 +95,7 @@ TEST(Win32UnicodeString, convertToUtf8) {
         char out[256];
         int len = Win32UnicodeString::convertToUtf8(out, sizeof(out),
                                                     kData[n].utf16);
-        EXPECT_EQ(strlen(kData[n].utf8) + 1, len);
+        EXPECT_EQ(strlen(kData[n].utf8) + 1U, (size_t)len);
         EXPECT_STREQ(kData[n].utf8, out);
 
         len = Win32UnicodeString::convertToUtf8(out, sizeof(out),
